+++ packages/playground/alias/vite.config.js
@@ -22,5 +22,9 @@ module.exports = {
   },
   build: {
     minify: false
+  },
+  define: {
+    __VUE_OPTIONS_API__: true,
+    __VUE_PROD_DEVTOOLS__: true
   }
 }