+++ packages/vite/src/node/ssr/ssrTransform.ts
@@ -25,7 +25,8 @@ export const ssrImportMetaKey = `__vite_ssr_import_meta__`
 
 export async function ssrTransform(
   code: string,
-  inMap: SourceMap | null
+  inMap: SourceMap | null,
+  url: string
 ): Promise<TransformResult | null> {
   const s = new MagicString(code)
 
@@ -38,6 +39,7 @@ export async function ssrTransform(
   let uid = 0
   const deps = new Set<string>()
   const idToImportMap = new Map<string, string>()
+  const declaredConst = new Set<string>()
 
   function defineImport(node: Node, source: string) {
     deps.add(source)
@@ -52,7 +54,7 @@ export async function ssrTransform(
   function defineExport(name: string, local = name) {
     s.append(
       `\nObject.defineProperty(${ssrModuleExportsKey}, "${name}", ` +
-        `{ get(){ return ${local} }})`
+        `{ enumerable: true, configurable: true, get(){ return ${local} }})`
     )
   }
 
@@ -153,6 +155,16 @@ export async function ssrTransform(
         ) {
           s.appendLeft(id.end, `: ${binding}`)
         }
+      } else if (
+        parent.type === 'ClassDeclaration' &&
+        id === parent.superClass
+      ) {
+        if (!declaredConst.has(id.name)) {
+          declaredConst.add(id.name)
+          // locate the top-most node containing the class declaration
+          const topNode = parentStack[1]
+          s.prependRight(topNode.start, `const ${id.name} = ${binding};\n`)
+        }
       } else {
         s.overwrite(id.start, id.end, binding)
       }
@@ -172,6 +184,9 @@ export async function ssrTransform(
       sources: inMap.sources,
       sourcesContent: inMap.sourcesContent
     }) as SourceMap
+  } else {
+    map.sources = [url]
+    map.sourcesContent = [code]
   }
 
   return {
@@ -208,11 +223,11 @@ function walk(
 
   ;(eswalk as any)(root, {
     enter(node: Node, parent: Node | null) {
-      parent && parentStack.push(parent)
-
       if (node.type === 'ImportDeclaration') {
         return this.skip()
       }
+
+      parent && parentStack.push(parent)
 
       if (node.type === 'MetaProperty' && node.meta.name === 'import') {
         onImportMeta(node)
@@ -301,6 +316,11 @@ function isRefIdentifier(id: Identifier, parent: _Node, parentStack: _Node[]) {
     if (parent.params.includes(id)) {
       return false
     }
+  }
+
+  // class method name
+  if (parent.type === 'MethodDefinition') {
+    return false
   }
 
   // property key